TasksPage: cleanup. Move functions to own helpers dir
