+++ app/pages/lab-pages-editor/helpers/constants.js
@@ -1,0 +1,2 @@
+export const TASK_KEY_PREFIX = 'T';
+export const STEP_KEY_PREFIX = 'P'; // Steps are known as Pages to users
