+++ app/pages/lab-pages-editor/helpers/createStep.js
@@ -1,0 +1,7 @@
+/*
+Creates a Step, with tasks, that's ready to be inserted into a Workflow.
+ */
+
+export default function createStep(stepKey, taskKeys = []) {
+  return [stepKey, { taskKeys }];
+}
